+++ examples/codes.rs
@@ -5,8 +5,8 @@ use escpos::{driver::*, errors::Result};
 fn main() -> Result<()> {
     env_logger::init();
 
-    let driver = NetworkDriver::open("192.168.1.248", 9100)?;
-    // let driver = ConsoleDriver::open(true);
+    // let driver = NetworkDriver::open("192.168.1.248", 9100)?;
+    let driver = ConsoleDriver::open(true);
     let mut printer = Printer::new(driver, Protocol::default());
     printer
         .debug_mode(Some(DebugMode::Dec))